+++ CHANGELOG.md
@@ -5,3 +5,4 @@
  - Make ParseData public
  - Make `PacketPool` and `(p *PacketPool) Add()` public
  - Add all stream types values
+ - Copy data into `d.originalBytes` in `parseDescriptors`, as we cannot count on it persisting